requirements.txt
alembic==1.7.7
attrs==21.4.0
black==22.3.0
Cerberus==1.3.4
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
flake8==4.0.1
Flask==2.1.2
greenlet==1.1.2
importlib-metadata==4.11.3
iniconfig==1.1.1
iniherit==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.960
mypy-extensions==0.4.3
mysqlclient==2.1.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.8
PyPika==0.48.9
pytest==7.1.2
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.36
tabulate==0.8.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.2.0
Werkzeug==2.1.2
yoyo-database-migrations==0.1.0
zipp==3.8.0
